moved logic to recognizer registry provider
