+++ presidio-analyzer/presidio_analyzer/analyzer_engine_provider.py
@@ -6,7 +6,8 @@ from typing import Optional, Union, List, Tuple, Dict, Any
 
 from presidio_analyzer import AnalyzerEngine, RecognizerRegistry, EntityRecognizer, PatternRecognizer
 from presidio_analyzer.nlp_engine import NlpEngineProvider, NlpEngine
-import presidio_analyzer.recognizer_registry
+from presidio_analyzer.recognizer_registry import RecognizerRegistry
+from presidio_analyzer.recognizer_registry_provider import RecognizerRegistryProvider
 
 logger = logging.getLogger("presidio-analyzer")
 
@@ -57,52 +58,6 @@ class AnalyzerEngineProvider:
         )
 
         return analyzer
-    
-    def _is_enabled(self, recognizer: Union[Dict[str, Any], str]) -> bool:
-        return not "enabled" in recognizer or recognizer["enabled"]
-
-    def _get_name(self, recognizer: Union[Dict[str, Any], str]) -> str:
-        if isinstance(recognizer, str):
-            return recognizer
-        return recognizer["name"]
-
-    def _get_items(self, recognizer: Union[Dict[str, Any], str]) -> Dict[str, Any]:
-        if isinstance(recognizer, str):
-            return {}
-        return recognizer.items()
-
-    def _get_languages(self, recognizer: Union[Dict[str, Any], str]) -> List[Dict[str, Any]]:
-        if isinstance(recognizer, str) or "supported_languages" not in recognizer:
-            logger.warning("language was not specified, defaulting to English")
-            return [{"supported_language": "en", "context": None}]
-        if isinstance(recognizer["supported_languages"][0], str):
-            return [{"supported_language": language, "context": None} for language in recognizer["supported_languages"]]
-        return [{"supported_language": language["language"], "context": language["context"]} for language in recognizer["supported_languages"]]
-
-    def _split_recognizers(self, recognizers: Union[Dict[str, Any], str]) -> Tuple[List[Dict[str, Any]],List[Dict[str, Any]]]:
-        """
-        splits the recognizer list to predefined and custom
-        """
-        predefined =[]
-        custom =[]
-
-        predefined = [recognizer for recognizer in recognizers if isinstance(recognizer, str) or ("type" in recognizer and recognizer["type"] == "predefined")]
-        # all recognizers are custom by default, type: custom can be mentioned as well
-        custom = [recognizer for recognizer in recognizers if not isinstance(recognizer, str) and ("type" not in recognizer or recognizer["type"] == "custom")]
-        return predefined, custom
-
-    def _create_recognizers(self, recognizer: Dict) -> List[EntityRecognizer]:
-        # in case supported_languages is not present, use the previous interface for custom recognizers
-        if not "supported_languages" in recognizer:
-            return [PatternRecognizer.from_dict(recognizer)]
-
-        recognizers = []
-
-        for supported_language in self._get_languages(recognizer):
-            copied_recognizer = {k: v for k, v in recognizer.items() if k not in ["enabled", "type", "supported_languages"]}
-            recognizers.append(PatternRecognizer.from_dict({**copied_recognizer, **supported_language}))
-
-        return recognizers
 
     def _load_recognizer_registry(self) -> RecognizerRegistry:
         if "recognizer_registry" not in self.configuration:
@@ -110,30 +65,9 @@ class AnalyzerEngineProvider:
                 "configuration file is missing 'recognizer_registry'. Using default configuration for recognizer registry"
             )
             return None
-        recognizer_registry = self.configuration["recognizer_registry"]
-        if isinstance(recognizer_registry, str):
-            if recognizer_registry!="predefined":
-                logger.warning(
-                "recognizer_registry contains unidentified value. Using default configuration for recognizer registry"
-                )
-            return None
-
-        recognizers=recognizer_registry["recognizers"]
-        recognizer_instances = []
-        predefined, custom = self._split_recognizers(recognizers)
-        for recognizer in predefined:
-            for language in self._get_languages(recognizer):
-                if self._is_enabled(recognizer):
-                    copied_recognizer = {k: v for k, v in self._get_items(recognizer) if k not in ["enabled", "type", "supported_languages", "name"]}
-                    recognizer_instances.append(getattr(presidio_analyzer.predefined_recognizers, self._get_name(recognizer), None)(**{**copied_recognizer, **language}))
-
-        for recognizer in custom:
-            if self._is_enabled(recognizer):
-                recognizer_instances.append(self._create_recognizers(recognizer))
-
-        global_regex_flags = recognizer_registry.get("global_regex_flags", None)
-
-        return RecognizerRegistry(recognizers=recognizer_instances, global_regex_flags=global_regex_flags)
+        registry_configuration = self.configuration["recognizer_registry"]
+        provider = RecognizerRegistryProvider(registry_configuration=registry_configuration)
+        return provider.create_engine_recognizer_registry()
     
     def _load_nlp_engine(self) -> NlpEngine:
         if "nlp_configuration" not in self.configuration:
